graph and node class set up
