+++ model/testing.py
@@ -1,5 +1,7 @@
-#from graphutils import create_random_graph
+# import sys
+# sys.path.insert(1, 'utils')
+# from graphutils import create_random_graph, print_graph
 
-#random_graph = create_random_graph(num_nodes = 5, num_edges = 15)
+# random_graph = create_random_graph(num_nodes = 5, num_edges = 15)
 
-print("hello world")
+# print_graph(random_graph,1)
